moving query to own module
