+++ src/oaklib/cli.py
@@ -7,12 +7,9 @@ Executed using "runoak" command
 
 # TODO: order commands.
 # See https://stackoverflow.com/questions/47972638/how-can-i-define-the-order-of-click-sub-commands-in-help
-import itertools
 import json
 import logging
 import os
-import re
-import secrets
 import statistics as stats
 import sys
 from collections import defaultdict
@@ -22,17 +19,11 @@ from pathlib import Path
 from time import time
 from types import ModuleType
 from typing import (
-    IO,
     Any,
-    Dict,
-    Iterable,
-    Iterator,
     List,
     Optional,
     TextIO,
-    Tuple,
     Type,
-    Union,
 )
 
 import click
@@ -64,7 +55,6 @@ from oaklib.datamodels.obograph import (
     Meta,
     PrefixDeclaration,
 )
-from oaklib.datamodels.search import create_search_configuration
 from oaklib.datamodels.settings import Settings
 from oaklib.datamodels.summary_statistics_datamodel import (
     GroupedStatistics,
@@ -73,17 +63,12 @@ from oaklib.datamodels.summary_statistics_datamodel import (
 from oaklib.datamodels.text_annotator import TextAnnotationConfiguration
 from oaklib.datamodels.validation_datamodel import ValidationConfiguration
 from oaklib.datamodels.vocabulary import (
-    DEVELOPS_FROM,
-    EQUIVALENT_CLASS,
     HAS_OBO_NAMESPACE,
     IS_A,
     IS_DEFINED_BY,
     OBSOLETION_RELATIONSHIP_PREDICATES,
-    OWL_CLASS,
-    OWL_OBJECT_PROPERTY,
     PART_OF,
     PREFIX_PREDICATE,
-    RDF_TYPE,
 )
 from oaklib.implementations.aggregator.aggregator_implementation import (
     AggregatorImplementation,
@@ -143,13 +128,14 @@ from oaklib.io.streaming_writer import StreamingWriter
 from oaklib.io.streaming_yaml_writer import StreamingYamlWriter
 from oaklib.mappers.ontology_metadata_mapper import OntologyMetadataMapper
 from oaklib.parsers.association_parser_factory import get_association_parser
+from oaklib.query import query_terms_iterator, curies_from_file, _process_predicates_arg
 from oaklib.resource import OntologyResource
 from oaklib.selector import get_adapter, get_resource_from_shorthand
 from oaklib.transformers.transformers_factory import (
     apply_ontology_transformation,
     get_ontology_transformer,
 )
-from oaklib.types import CURIE, PRED_CURIE
+from oaklib.types import CURIE
 from oaklib.utilities import table_filler
 from oaklib.utilities.apikey_manager import set_apikey_value
 from oaklib.utilities.associations.association_differ import AssociationDiffer
@@ -194,7 +180,6 @@ from oaklib.utilities.obograph_utils import (
 from oaklib.utilities.publication_utils.pubmed_wrapper import PubmedWrapper
 from oaklib.utilities.semsim.similarity_utils import load_information_content_map
 from oaklib.utilities.subsets.slimmer_utils import (
-    filter_redundant,
     roll_up_to_named_subset,
 )
 from oaklib.utilities.table_filler import ColumnDependency, TableFiller, TableMetadata
@@ -470,53 +455,6 @@ group_by_prefix_option = click.option(
 )
 
 
-def _process_predicates_arg(
-    predicates_str: str,
-    expected_number: Optional[int] = None,
-    exclude_predicates_str: Optional[str] = None,
-    impl: Optional[OntologyInterface] = None,
-) -> Optional[List[PRED_CURIE]]:
-    if predicates_str is None and exclude_predicates_str is None:
-        return None
-    if predicates_str is None:
-        inputs = []
-    elif "," in predicates_str:
-        inputs = predicates_str.split(",")
-    else:
-        inputs = predicates_str.split("+")
-    preds = [_shorthand_to_pred_curie(p) for p in inputs]
-    if exclude_predicates_str:
-        if "," in exclude_predicates_str:
-            exclude_inputs = exclude_predicates_str.split(",")
-        else:
-            exclude_inputs = exclude_predicates_str.split("+")
-        exclude_preds = [_shorthand_to_pred_curie(p) for p in exclude_inputs]
-        if not preds:
-            if not impl or not isinstance(impl, BasicOntologyInterface):
-                raise ValueError("Must provide an BasicOntologyInterface to exclude predicates")
-            preds = list(impl.entities(owl_type=OWL_OBJECT_PROPERTY))
-        preds = [p for p in preds if p not in exclude_preds]
-        logging.info(f"Excluding predicates: {exclude_preds} yields: {preds}")
-    if expected_number and len(preds) != expected_number:
-        raise ValueError(f"Expected {expected_number} parses of {predicates_str}, got: {preds}")
-    return preds
-
-
-# TODO: move to vocab
-def _shorthand_to_pred_curie(shorthand: str) -> PRED_CURIE:
-    if shorthand == "i":
-        return IS_A
-    elif shorthand == "p":
-        return PART_OF
-    elif shorthand == "d":
-        return DEVELOPS_FROM
-    elif shorthand == "t":
-        return RDF_TYPE
-    elif shorthand == "e":
-        return EQUIVALENT_CLASS
-    else:
-        return shorthand
-
 
 def _skip_if_absent(if_absent: bool, v: Any):
     if if_absent:
@@ -557,307 +495,6 @@ def _apply_changes(impl, changes: List[kgcl.Change]):
         impl.save()
 
 
-# A list whose members are either strings (search terms, curies, or directives)
-# or nested lists.
-# TODO: Replace this with an explicit query model with boolean operations
-NESTED_LIST = Union[List[str], List["NESTED_LIST"]]
-
-
-def nest_list_of_terms(terms: List[str]) -> NESTED_LIST:
-    """
-    Gives a list of terms (typically passed on command line),
-    replace blocks between '[', ..., ']' with nested lists of the contents
-
-    :param terms:
-    :return:
-    """
-    nested, rest = _nest_list_of_terms(terms)
-    if rest:
-        raise ValueError(f"Unparsed: {rest}")
-    return nested
-
-
-def _nest_list_of_terms(terms: List[str]) -> Tuple[NESTED_LIST, List[str]]:
-    nested = []
-    while len(terms) > 0:
-        term = terms[0]
-        terms = terms[1:]
-        if term == "[":
-            nxt, rest = _nest_list_of_terms(terms)
-            terms = rest
-            nested.append(nxt)
-        elif term == "]":
-            return nested, terms
-        else:
-            nested.append(term)
-    return nested, []
-
-
-def curies_from_file(
-    file: IO, adapter: Optional[BasicOntologyInterface] = None, allow_labels=False, strict=False
-) -> Iterator[CURIE]:
-    """
-    yield an iterator over CURIEs by parsing a file.
-
-    The file can contain any content, so long as each line
-    starts with a CURIE followed by whitespace -- the remainder of the line
-    is ignored
-
-    :param file:
-    :param adapter: if provided, will be used to resolve CURIEs
-    :param allow_labels: if true, will allow inputs to be labels
-    :param strict: if true, will raise an error if a CURIE cannot be resolved
-    :return:
-    """
-    line_no = 0
-    if allow_labels and not adapter:
-        raise ValueError("Must provide an adapter to resolve labels")
-    for line in file.readlines():
-        line_no += 1
-        if ":" in line or not allow_labels:
-            m = re.match(r"^(\S+)", line)
-            curie = m.group(1)
-            if curie == "id" and line_no == 1:
-                continue
-            yield curie
-        elif allow_labels:
-            candidates = adapter.curies_by_label(line.strip())
-            if strict and len(candidates) != 1:
-                raise ValueError(
-                    f"Could not resolve label {line} to a single CURIE, got {candidates}"
-                )
-            yield from candidates
-
-
-def query_terms_iterator(query_terms: NESTED_LIST, impl: BasicOntologyInterface) -> Iterator[CURIE]:
-    """
-    Turn list of tokens that represent a term query into an iterator for curies.
-
-    For examples, see test_cli
-
-    :param query_terms:
-    :param impl:
-    :return:
-    """
-    # TODO: reimplement using an explicit query model
-    results: Iterable[CURIE] = iter([])
-    predicates = None
-    if isinstance(query_terms, tuple):
-        query_terms = list(query_terms)
-
-    def _parse_params(s: str) -> Dict:
-        # some query terms are parameterized using the syntax
-        # .<TOKEN>//<P1>=<V1>//<P2>=<V2>...
-        d = {}
-        m = re.match(r"\.\w+//(.+)", s)
-        if m:
-            for p in m.group(1).split("//"):
-                if "=" not in p:
-                    raise ValueError(f"All arguments must be of param=val form, got {p} in {s}")
-                [k, v] = p.split("=")
-                if k == "p":
-                    k = "predicates"
-                if k == "predicates":
-                    v = _process_predicates_arg(v)
-                if k == "prefixes":
-                    v = v.split(",")
-                d[k] = v
-        return d
-
-    def chain_results(v):
-        if isinstance(v, str):
-            v = iter([v])
-        nonlocal results
-        results = itertools.chain(results, v)
-
-    # queries can be nested using square brackets
-    query_terms = nest_list_of_terms(query_terms)
-
-    while len(query_terms) > 0:
-        # process each query term. A query term is either:
-        # 1. local, in which case is appends to the result iterator
-        # 2. global, where it applies all subsequent terms
-        term = query_terms[0]
-        query_terms = query_terms[1:]
-        if term == "-":
-            # read from stdin
-            chain_results(curies_from_file(sys.stdin))
-        elif isinstance(term, list):
-            chain_results(query_terms_iterator(term, impl))
-        elif term.startswith(".load="):
-            # load a file of IDs
-            fn = term.replace(".load=", "")
-            with open(fn) as file:
-                chain_results(curies_from_file(file))
-        elif term.startswith(".idfile"):
-            # load a file of IDs
-            fn = query_terms.pop(0)
-            logging.info(f"Reading ids from {fn}")
-            file = open(fn)
-            chain_results(curies_from_file(file))
-        elif term.startswith(".termfile"):
-            # load a file of queries
-            fn = query_terms.pop(0)
-            with open(fn) as file:
-                lines = [line.strip() for line in file.readlines()]
-                query_terms = lines + query_terms
-        elif re.match(r"^([\w\-\.]+):(\S+)$", term):
-            # CURIE
-            chain_results(term)
-        elif re.match(r"^http(\S+)$", term):
-            # URI
-            chain_results(term)
-        elif re.match(r"^\.predicates=(\S*)$", term):
-            logging.warning("Deprecated: pass as parameter instead")
-            m = re.match(r"^\.predicates=(\S*)$", term)
-            predicates = _process_predicates_arg(m.group(1))
-        elif term == ".and":
-            # boolean term: consume the result of the query and intersect
-            rest = list(query_terms_iterator(query_terms, impl))
-            for x in results:
-                if x in rest:
-                    yield x
-            query_terms = []
-        elif term == ".xor":
-            # boolean term: consume the result of the query and xor
-            rest = list(query_terms_iterator(query_terms, impl))
-            remaining = []
-            for x in results:
-                if x not in rest:
-                    yield x
-                else:
-                    remaining.append(x)
-            for x in rest:
-                if x not in remaining:
-                    yield x
-            query_terms = []
-        elif term == ".not" or term == ".minus":
-            # boolean term: consume the result of the query and subtract
-            rest = list(query_terms_iterator(query_terms, impl))
-            for x in results:
-                if x not in rest:
-                    yield x
-            query_terms = []
-        elif term == ".or":
-            # or is implicit
-            pass
-        elif term.startswith(".all"):
-            chain_results(impl.entities(filter_obsoletes=False))
-        elif term.startswith(".classes"):
-            chain_results(impl.entities(owl_type=OWL_CLASS))
-        elif term.startswith(".relations"):
-            chain_results(impl.entities(owl_type=OWL_OBJECT_PROPERTY))
-        elif term.startswith(".rand"):
-            params = _parse_params(term)
-            sample_size = params.get("n", "100")
-            entities = list(impl.entities())
-            sample = [
-                entities[secrets.randbelow(len(entities))] for x in range(1, int(sample_size))
-            ]
-            chain_results(sample)
-        elif term.startswith(".in"):
-            # subset query
-            subset = query_terms[0]
-            query_terms = query_terms[1:]
-            chain_results(impl.subset_members(subset))
-        elif term.startswith(".is_obsolete"):
-            chain_results(impl.obsoletes())
-        elif term.startswith(".non_obsolete"):
-            chain_results(impl.entities(filter_obsoletes=True))
-        elif term.startswith(".dangling"):
-            chain_results(impl.dangling())
-        elif term.startswith(".filter"):
-            # arbitrary python expression
-            expr = query_terms[0]
-            query_terms = query_terms[1:]
-            chain_results(eval(expr, {"impl": impl, "terms": results}))  # noqa
-        elif term.startswith(".query"):
-            # arbitrary SPARQL query
-            params = _parse_params(term)
-            prefixes = params.get("prefixes", None)
-            query = query_terms[0]
-            query_terms = query_terms[1:]
-            chain_results([list(v.values())[0] for v in impl.query(query, prefixes=prefixes)])
-        elif term.startswith(".desc"):
-            # graph query: descendants
-            params = _parse_params(term)
-            this_predicates = params.get("predicates", predicates)
-            rest = list(query_terms_iterator([query_terms[0]], impl))
-            query_terms = query_terms[1:]
-            if isinstance(impl, OboGraphInterface):
-                chain_results(impl.descendants(rest, predicates=this_predicates))
-            else:
-                raise NotImplementedError
-        elif term.startswith(".child"):
-            # graph query: children
-            params = _parse_params(term)
-            this_predicates = params.get("predicates", predicates)
-            rest = list(query_terms_iterator([query_terms[0]], impl))
-            query_terms = query_terms[1:]
-            children = [
-                s for s, _p, _o in impl.relationships(objects=rest, predicates=this_predicates)
-            ]
-            chain_results(children)
-        elif term.startswith(".parent"):
-            # graph query: parents
-            params = _parse_params(term)
-            this_predicates = params.get("predicates", predicates)
-            rest = list(query_terms_iterator([query_terms[0]], impl))
-            query_terms = query_terms[1:]
-            parents = [
-                o for _s, _p, o in impl.relationships(subjects=rest, predicates=this_predicates)
-            ]
-            chain_results(parents)
-        elif term.startswith(".sib"):
-            # graph query: siblings
-            params = _parse_params(term)
-            this_predicates = params.get("predicates", predicates)
-            rest = list(query_terms_iterator([query_terms[0]], impl))
-            query_terms = query_terms[1:]
-            parents = [
-                o for _s, _p, o in impl.relationships(subjects=rest, predicates=this_predicates)
-            ]
-            sibs = [
-                s for s, _p, _o in impl.relationships(objects=parents, predicates=this_predicates)
-            ]
-            chain_results(sibs)
-        elif term.startswith(".anc"):
-            # graph query: ancestors
-            params = _parse_params(term)
-            this_predicates = params.get("predicates", predicates)
-            rest = list(query_terms_iterator([query_terms[0]], impl))
-            query_terms = query_terms[1:]
-            if isinstance(impl, OboGraphInterface):
-                chain_results(impl.ancestors(rest, predicates=this_predicates))
-            else:
-                raise NotImplementedError
-        elif term.startswith(".mrca"):
-            # graph query: most recent common ancestors
-            params = _parse_params(term)
-            this_predicates = params.get("predicates", predicates)
-            rest = list(query_terms_iterator([query_terms[0]], impl))
-            query_terms = query_terms[1:]
-            if isinstance(impl, SemanticSimilarityInterface):
-                chain_results(
-                    impl.setwise_most_recent_common_ancestors(rest, predicates=this_predicates)
-                )
-            else:
-                raise NotImplementedError
-        elif term.startswith(".nr"):
-            # graph query: non-redundant
-            params = _parse_params(term)
-            this_predicates = params.get("predicates", predicates)
-            rest = list(query_terms_iterator([query_terms[0]], impl))
-            query_terms = query_terms[1:]
-            chain_results(filter_redundant(impl, rest, this_predicates))
-        else:
-            # term is not query syntax: feed directly to search
-            if not isinstance(impl, SearchInterface):
-                raise NotImplementedError(f"Search not implemented for {type(impl)}")
-            cfg = create_search_configuration(term)
-            logging.info(f"Search config: {term} => {cfg}")
-            chain_results(impl.basic_search(cfg.search_terms[0], config=cfg))
-    yield from results
 
 
 @click.group()
@@ -1216,7 +853,7 @@ def obsoletes(
 
     More examples:
 
-       https://github.com/INCATools/ontology-access-kit/blob/main/notebooks/Commands/TaxonConstraints.ipynb
+       https://github.com/INCATools/ontology-access-kit/blob/main/notebooks/Commands/Obsoletes.ipynb
 
     Python API:
 
@@ -2334,7 +1971,7 @@ def paths(
         ix = terms.index("@")
         logging.info(f"Splitting terms {terms} on {ix}")
         start_curies = list(query_terms_iterator(terms[0:ix], impl))
-        end_curies = list(query_terms_iterator(terms[ix + 1 :], impl))
+        end_curies = list(query_terms_iterator(terms[ix + 1:], impl))
         all_curies = start_curies + end_curies
     else:
         start_curies = list(query_terms_iterator(terms, impl))
@@ -2968,7 +2605,7 @@ def similarity(
         ix = terms.index("@")
         logging.info(f"Splitting terms {terms} on {ix}")
         set1it = query_terms_iterator(terms[0:ix], impl)
-        set2it = query_terms_iterator(terms[ix + 1 :], impl)
+        set2it = query_terms_iterator(terms[ix + 1:], impl)
     else:
         if set1_file:
             logging.info(f"Getting set1 from {set1_file}")
@@ -3054,7 +2691,7 @@ def termset_similarity(
     terms = list(terms)
     ix = terms.index("@")
     set1 = list(query_terms_iterator(terms[0:ix], impl))
-    set2 = list(query_terms_iterator(terms[ix + 1 :], impl))
+    set2 = list(query_terms_iterator(terms[ix + 1:], impl))
     logging.info(f"Set1={set1}")
     logging.info(f"Set2={set2}")
     actual_predicates = _process_predicates_arg(predicates)
@@ -3865,7 +3502,7 @@ def query(query, autolabel: bool, output: str, output_type: str, prefixes: str, 
 
     # batch query results: this allows us to produce buffered output
     # for a long-running query
-    for row_it in chunk(impl.query(query, prefixes=prefixes)):
+    for row_it in chunk(impl.onto_query(query, prefixes=prefixes)):
         rows = []
         # collect rows in batch, and set label_fields if not already set
         for r in row_it:
@@ -4101,14 +3738,18 @@ def mappings(terms, maps_to_source, autolabel: bool, output, output_type, mapper
     mapper_impl = impl
     if mapper:
         mapper_impl = get_adapter(mapper)
+
+    def _emit(mapping):
+        if autolabel:
+            impl.inject_mapping_labels([mapping])
+        writer.emit(mapping)
+
     if len(terms) == 0:
         logging.info(f"No terms provided: fetching all mappings for {maps_to_source}")
         for mapping in mapper_impl.sssom_mappings_by_source(
             subject_or_object_source=maps_to_source
         ):
-            if autolabel:
-                impl.inject_mapping_labels([mapping])
-            writer.emit(mapping)
+            _emit(mapping)
     else:
         logging.info(f"Fetching mappings for {terms}")
         for curie_it in chunk(query_terms_iterator(terms, impl)):
@@ -4117,9 +3758,7 @@ def mappings(terms, maps_to_source, autolabel: bool, output, output_type, mapper
                     f"{maps_to_source.lower()}:"
                 ):
                     continue
-                if autolabel:
-                    impl.inject_mapping_labels([mapping])
-                writer.emit(mapping)
+                _emit(mapping)
     writer.finish()
 
 
@@ -4881,7 +4520,7 @@ def associations_matrix(
         ix = terms.index("@")
         logging.info(f"Splitting terms into two, position = {ix}")
         curies1 = list(query_terms_iterator(terms[0:ix], impl))
-        curies2 = list(query_terms_iterator(terms[ix + 1 :], impl))
+        curies2 = list(query_terms_iterator(terms[ix + 1:], impl))
     else:
         curies1 = list(query_terms_iterator(terms, impl))
         curies2 = list(curies1)
@@ -5788,10 +5427,16 @@ def cache_clear(days_old: int):
     show_default=True,
     help="Clean prefix map and mappings before generating an output.",
 )
+@click.option(
+    "--exclude-mapped/--no-exclude-mapped",
+    default=False,
+    show_default=True,
+    help="Return only mappings for subjects that have not been mapped",
+)
 @output_option
 @click.argument("terms", nargs=-1)
 def lexmatch(
-    output, recreate, ensure_strict_prefixes, rules_file, lexical_index_file, add_labels, terms
+    output, recreate, ensure_strict_prefixes, rules_file, lexical_index_file, add_labels, terms,  exclude_mapped,
 ):
     """
     Performs lexical matching between pairs of terms in one more more ontologies.
@@ -5861,7 +5506,7 @@ def lexmatch(
             ix = terms.index("@")
             logging.info(f"Splitting terms into two, position = {ix}")
             subjects = list(query_terms_iterator(terms[0:ix], impl))
-            objects = list(query_terms_iterator(terms[ix + 1 :], impl))
+            objects = list(query_terms_iterator(terms[ix + 1:], impl))
         else:
             subjects = list(query_terms_iterator(terms, impl))
             objects = subjects
@@ -5879,6 +5524,7 @@ def lexmatch(
             syn_rules = [x.synonymizer for x in ruleset.rules if x.synonymizer]
         else:
             syn_rules = []
+        logging.info(f"Synonymizer rules: {syn_rules}")
         ix = create_lexical_index(impl, synonym_rules=syn_rules)
     if lexical_index_file:
         if recreate:
@@ -5895,6 +5541,19 @@ def lexmatch(
         prefix_map=prefix_map,
         ensure_strict_prefixes=ensure_strict_prefixes,
     )
+    # TODO: coordinate with sssom-py to have simpler ways of filtering, no reliance on pandas
+    if exclude_mapped and not msdf.df.empty:
+        if not isinstance(impl, MappingProviderInterface):
+            raise ValueError(f"{type(impl)} does not implement MappingProviderInterface")
+        df = msdf.df
+        subjects = set(df["subject_id"])
+        existing_mappings = list(impl.sssom_mappings(subjects))
+        pairs = [(m.subject_id, m.object_source) for m in existing_mappings]
+        pairs += [(m.object_id, m.subject_source) for m in existing_mappings]
+        # filter df by pairs
+        df = df[df.apply(lambda x: (x["subject_id"], x["object_id"].split(":")[0]) not in pairs, axis=1)]
+        df = df[df.apply(lambda x: (x["object_id"], x["subject_id"].split(":")[0]) not in pairs, axis=1)]
+        msdf.df = df
     sssom_writers.write_table(msdf, output)
